requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.5.0+cpu
torchvision==0.6.0+cpu
Pillow==7.1.2
scikit-image
numpy
matplotlib
IPython
flask
flask_restful
werkzeug
